pyproject.toml
[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[project]
name = "sysdweb"
version = "1.1.6"
description = "Control systemd services through Web or REST API"
readme = "README.md"
requires-python = ">=3.10"
license = {file = "LICENSE"}
keywords = ["REST API", "systemd"]
authors = [
  {name = "Óscar García Amor", email = "ogarcia@connectical.com"}
]
classifiers = [
  "Development Status :: 5 - Production/Stable",
  "Environment :: Web Environment",
  "Framework :: Bottle",
  "Intended Audience :: End Users/Desktop",
  "Intended Audience :: System Administrators",
  "License :: OSI Approved :: GNU General Public License v3 (GPLv3)",
  "Natural Language :: English",
  "Operating System :: POSIX :: Linux",
  "Programming Language :: Python :: 3",
  "Topic :: System",
  "Topic :: Utilities"
]
dependencies = [
  "bottle ~= 0.13",
  "dbus-python ~= 1.4",
  "python-pam ~= 2.0.2",
  "six ~= 1.17",
  "systemd-python"
]

[tool.setuptools.packages.find]
include = ["sysdweb*"]

[tool.setuptools.package-data]
"*" = [
  "sysdweb.conf"
]
sysdweb = [
  "templates/static/css/*",
  "templates/static/img/*",
  "templates/static/js/*",
  "templates/views/*"
]

[project.scripts]
sysdweb = "sysdweb.main:main"

[project.urls]
"Bug Reports" = "https://github.com/ogarcia/sysdweb/issues"
"Source" = "https://github.com/ogarcia/sysdweb"
